requirements.txt
piexif==1.0.8
pillow==3.4.1
